Updated squidguard

+++ proxy_config/update_proxy_services.sh
@@ -5,14 +5,19 @@ cd ~
 sudo cp -f ./piproxy/proxy_config/proxy.pac /usr/share/nginx/html/proxy.pac
 sudo chmod 0755 /usr/share/nginx/html/proxy.pac
 
-# Squid Proxy and Guard Config File
+# Squid Proxy Config File
 sudo cp -f ./piproxy/proxy_config/squid.conf /etc/squid/squid.conf
-sudo cp -f ./piproxy/proxy_config/squidGuard.conf /etc/squidguard/squidGuard.conf
 
-# Update squidguard post updates
-sudo squidGuard -d -b -P -C all
-sudo chown -R proxy:proxy /var/lib/squidguard/db/
-
+# Update squidguard post updates, this takes some time...
+if [ -e ./status/squidguard_updated ]
+then
+  echo "SquidGuard update not needed, skipping..."
+else
+  sudo cp -f ./piproxy/proxy_config/squidGuard.conf /etc/squidguard/squidGuard.conf
+  sudo squidGuard -d -b -P -C all
+  sudo chown -R proxy:proxy /var/lib/squidguard/db/
+  touch ./status/squidguard_updated
+fi
 
 # nginx Config File
